+++ src/main/java/min/taskflow/task/controllder/TaskController.java
@@ -7,7 +7,7 @@ import min.taskflow.task.dto.condition.TaskSearchCondition;
 import min.taskflow.task.dto.request.StatusUpdateRequest;
 import min.taskflow.task.dto.request.TaskCreateRequest;
 import min.taskflow.task.dto.request.TaskUpdateRequest;
-import min.taskflow.task.dto.response.TaskResponse;
+import min.taskflow.task.dto.response.task.TaskResponse;
 import min.taskflow.task.service.commandService.ExternalCommandTaskService;
 import min.taskflow.task.service.queryService.ExternalQueryTaskService;
 import org.springframework.data.domain.Page;
