+++ src/test/java/min/taskflow/task/TaskServiceTest.java
@@ -1,7 +1,7 @@
 package min.taskflow.task;
 
 import min.taskflow.task.dto.request.TaskCreateRequest;
-import min.taskflow.task.dto.response.TaskResponse;
+import min.taskflow.task.dto.response.task.TaskResponse;
 import min.taskflow.task.entity.Priority;
 import min.taskflow.task.entity.Task;
 import min.taskflow.task.mapper.TaskMapper;
